move HumanInputStream to test library for reusability
